Extract poll answer process
